# src/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.0'
  title: Care Platform
  description: 'This is the API definition for the Care Platform.

  Authentication  Nodes:
  Click Authorize to enter User and Component authorization codes.

  - userAuth: The JWT token provided by the login command. Since the swagger 2.0 (OpenAPI) specification does
       not natively support bearer authentication. Please use this APIKey field. Prefix it with the word Bearer,
       like this: Bearer JWT_TOKEN
  - componentAuth: The key of the component, this can be obtained by making GET requests to .../components.'
host: localhost:10010
basePath: /v0
schemes:
- http
#s
consumes:
- application/json
produces:
- application/json
securityDefinitions:
   userAuth:
     type: apiKey
     in: header
     name: X-API-Key

   componentAuth:
     type: apiKey
     in: header
     name: X-API-Key
paths:
...
  /users/me/clients/tentative:
      x-swagger-router-controller: clients
      get:
        security:
        - userAuth: []
        description: Get all clients that are you have been invited to add to your account.
        summary: Get the tentative clients associated with your account
        tags:
        - client
        operationId: getUsersClientsTentative
        produces:
        - application/
        parameters: []
        responses:
          200:
            description: OK
            schema:
              type: array
              items:
                $ref: '#/definitions/ClientWithID'
          401:
            description: You aren't authorised to make this request
  /users/me/clients:
    x-swagger-router-controller: clients
    get:
      security: 
      - userAuth: [] 
      description: Get all clients that are associated with your user account.
      summary: Get the clients associated with your account
      tags:
      - client
      operationId: getUsersClients
      produces:
      - application/
      parameters: []
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ClientWithID'
        401:
          description: You aren't authorised to make this request

    post:
      security: 
      - userAuth: [] 
      description: The new user created from the information provided will be initially  associated with your account.
      summary: Create a new client associated with your account
      tags:
      - client
      operationId: postUsersClients
      produces:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        description: New client details
        schema:
          $ref: '#/definitions/Client'
      responses:
        201:
          description: Client was successfully created
          schema:
            $ref: '#/definitions/ClientWithID'
        401:
          description: You aren't authorised to make this change
        406:
          description: The data provided is not correct
  /users/me/clients/{clientID}:
    x-swagger-router-controller: clients
    delete:
      security: 
      - userAuth: [] 
      description: Remove the link between your account and the client specified in  the clientID. If you are the last user for the client, the client will also be deleted. You will need to provide your user credentials for this request.
      summary: Remove client from user account
      tags:
      - client
      operationId: deleteUserClient
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: ID of client to remove from account
      - name: body
        in: body
        required: true
        description: Credentials Pair
        schema:
          $ref: '#/definitions/Credentials'
      responses:
        200:
          description: Deleted
        401:
          description: You aren't authorised to make this request
        404:
          description: There is no client with this ID or you do not have permission to access it.
  /clients/{clientID}:
    x-swagger-router-controller: clients
    get:
      security:
      - userAuth: []
      description: Get client information by ID passed in path
      summary: Get client information by ID
      tags:
      - client
      operationId: getClient
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: ID of client to get
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ClientWithID'
        401:
          description: You aren't authorised to make this request
        404:
          description: There is no client with this ID or you do not have permission to access it.
      
    put:
      security:
      - userAuth: []
      description: Update basic details of the client, such as their name.  Information for fields which should not be updated should not be  supplied.
      summary: Update the information associated with a client.
      tags:
      - client
      operationId: updateClient
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: ID of client to get
      - name: body
        in: body
        required: true
        description: New client details
        schema:
          $ref: '#/definitions/Client'
      responses:
        201:
          description: Client was successfully updated
          schema:
            $ref: '#/definitions/ClientWithID'
        401:
          description: You aren't authorised to make this change
        406:
          description: The data provided is not correct
          
    delete:
      security:
      - userAuth: []
      description: Delete client with ID passed in path. You will need to provide your user credentials with this request.
      summary: Delete client by ID
      tags:
      - client
      operationId: deleteClient
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: ID of client to delete
      - name: body
        in: body
        required: true
        description: Credentials Pair
        schema:
          $ref: '#/definitions/Credentials'
      responses:
        200:
          description: Deleted
        401:
          description: You aren't authorised to make this request
        404:
          description: There is no client with this ID or you do not have permission to access it.
  /clients/{clientID}/share:
    x-swagger-router-controller: clients
    post:
      security:
      - userAuth: []
      description: Specify an existing user's email address to send an invitation to for  them to manage the client.
      summary: Invite user to access the client's data
      tags:
      - client
      operationId: shareClient
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: ID of client to share
      - name: body
        in: body
        required: true
        description: User to share the client with
        schema:
          $ref: '#/definitions/Invitation'
      responses:
        200:
          description: User was invited to manage this client.
        401:
          description: Invalid credentials supplied
        404:
          description: The client or user does not exist.
  /clients/{clientID}/accept:
    x-swagger-router-controller: clients
    post:
      security:
      - userAuth: []
      description: Accept a client that has been shared with you
      summary:  Accept a client that has been shared with you
      tags:
      - client
      operationId: acceptClient
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: ID of client to accept
      - name: body
        in: body
        required: true
        description: No body required
        schema:
          $ref: '#/definitions/Empty'
      responses:
        200:
          description: The client was accepted
        401:
          description: Invalid credentials supplied
        404:
          description: The client does not exist or is not tentative
...
definitions:
...
  Name:
    title: Name
    type: object
    properties:
      title:
        example: Dr
        type: string
      firstNames:
        example: Grace
        type: string
      surnames:
        example: Hopper
        type: string
...
  Credentials:
    title: Credentials
    type: object
    properties:
      email:
        example: carer@example.com
        type: string
      password:
        example: password
        type: string
    required:
    - email
    - password
...
  ClientWithID:
    title: ClientWithID
    type: object
    properties:
      id:
        type: integer
      name:
        $ref: '#/definitions/Name'
  Client:
    title: Client
    type: object
    properties:
      name:
        $ref: '#/definitions/Name'
    required:
    - name
  Invitation:
    title: Invitation
    type: object
    properties:
      email:
        example: carer@example.com
        type: string
      admin:
        example: false
        type: boolean
    required:
    - email
    - admin
    #component definitions
...
  Empty:
      title: empty
      type: object
...